Fixed save as uint8

+++ utils.py
@@ -107,11 +107,12 @@ def save_pred_as_nifti(pred, save_dir, data_dir, subject_name):
     pred_for_nifti = pred[0] + pred[1]*2 + pred[2]*3
     pred_for_nifti = np.squeeze(pred_for_nifti)
     pred_for_nifti = reshape_input(pred_for_nifti)
+    pred_for_nifti = pred_for_nifti.astype(np.uint8)
 
     affine, header = fetch_affine_header(subject_name, data_dir)
     pred_nifti = nib.nifti1.Nifti1Image(pred_for_nifti, affine=affine, header=header)
     filename = f'{subject_name}.nii.gz'
-    nib.nifti1.save(pred_nifti.astype(np.uint8), os.path.join(save_dir, filename))
+    nib.nifti1.save(pred_nifti, os.path.join(save_dir, filename))
 
 def compute_loss(output, seg, loss_functs, loss_weights):
     # Compute weighted loss, summed across each region.
